# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POS Kafe Multi-Cabang API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /menu-categories:
    get:
      operationId: getMenuCategories
      tags: [menu]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of menu categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuCategory"
    post:
      operationId: createMenuCategory
      tags: [menu]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
  /menu-categories/{id}:
    patch:
      operationId: updateMenuCategory
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuCategoryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuCategory"
    delete:
      operationId: deleteMenuCategory
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MENU ITEMS
  /menu-items:
    get:
      operationId: getMenuItems
      tags: [menu]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
    post:
      operationId: createMenuItem
      tags: [menu]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemInput"
      responses:
        "201":
          description: Menu item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
  /menu-items/{id}:
    get:
      operationId: getMenuItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Menu item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    patch:
      operationId: updateMenuItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
    delete:
      operationId: deleteMenuItem
      tags: [menu]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TABLES
...
components:
  schemas:
...
    MenuCategory:
      type: object
      required: [id, name, branchId, sortOrder, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        branchId:
          type: integer
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    MenuCategoryInput:
      type: object
      required: [name, branchId]
      properties:
        name:
          type: string
        branchId:
          type: integer
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
    MenuCategoryUpdate:
      type: object
      properties:
        name:
          type: string
        iconUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean

    # MENU ITEM
    MenuItem:
      type: object
      required: [id, name, price, categoryId, branchId, isAvailable, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        branchId:
          type: integer
        station:
          type: ["string", "null"]
        preparationTime:
          type: ["integer", "null"]
        isAvailable:
          type: boolean
        createdAt:
          type: string
    MenuItemInput:
      type: object
      required: [name, price, categoryId, branchId]
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        branchId:
          type: integer
        station:
          type: ["string", "null"]
        preparationTime:
          type: ["integer", "null"]
    MenuItemUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        categoryId:
          type: integer
        station:
          type: ["string", "null"]
        preparationTime:
          type: ["integer", "null"]
        isAvailable:
          type: boolean

    # TABLE
...